+++ packages/expo-sqlite/src/NativeSession.ts
@@ -2,6 +2,7 @@
  * A type that represents a changeset.
  */
 export type Changeset = Uint8Array;
+type NativeChangeset = ArrayBuffer;
 
 export type SQLiteAnyDatabase = any;
 
@@ -12,13 +13,16 @@ export declare class NativeSession {
   public enableAsync(database: SQLiteAnyDatabase, enabled: boolean): Promise<void>;
   public closeAsync(database: SQLiteAnyDatabase): Promise<void>;
 
-  public createChangesetAsync(database: SQLiteAnyDatabase): Promise<Changeset>;
-  public createInvertedChangesetAsync(database: SQLiteAnyDatabase): Promise<Changeset>;
-  public applyChangesetAsync(database: SQLiteAnyDatabase, changeset: Changeset): Promise<void>;
+  public createChangesetAsync(database: SQLiteAnyDatabase): Promise<NativeChangeset>;
+  public createInvertedChangesetAsync(database: SQLiteAnyDatabase): Promise<NativeChangeset>;
+  public applyChangesetAsync(
+    database: SQLiteAnyDatabase,
+    changeset: NativeChangeset
+  ): Promise<void>;
   public invertChangesetAsync(
     database: SQLiteAnyDatabase,
-    changeset: Changeset
-  ): Promise<Changeset>;
+    changeset: NativeChangeset
+  ): Promise<NativeChangeset>;
 
   //#endregion
 
@@ -28,10 +32,13 @@ export declare class NativeSession {
   public enableSync(database: SQLiteAnyDatabase, enabled: boolean): void;
   public closeSync(database: SQLiteAnyDatabase): void;
 
-  public createChangesetSync(database: SQLiteAnyDatabase): Changeset;
-  public createInvertedChangesetSync(database: SQLiteAnyDatabase): Changeset;
-  public applyChangesetSync(database: SQLiteAnyDatabase, changeset: Changeset): void;
-  public invertChangesetSync(database: SQLiteAnyDatabase, changeset: Changeset): Changeset;
+  public createChangesetSync(database: SQLiteAnyDatabase): NativeChangeset;
+  public createInvertedChangesetSync(database: SQLiteAnyDatabase): NativeChangeset;
+  public applyChangesetSync(database: SQLiteAnyDatabase, changeset: NativeChangeset): void;
+  public invertChangesetSync(
+    database: SQLiteAnyDatabase,
+    changeset: NativeChangeset
+  ): NativeChangeset;
 
   //#endregion
 }